requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
black==24.10.0
cchardet==2.1.7
certifi==2024.8.30
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.4.0
click==8.1.7
color-matcher==0.5.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
ddt==1.7.2
Deprecated==1.2.14
docutils==0.21.2
einops==0.8.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
Jinja2==3.1.3
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
lightning-utilities==0.11.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matrix-client==0.4.0
mdurl==0.1.2
mpmath==1.3.0
mss==9.0.2
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
onnxruntime-gpu==1.20.0
opencv-python==4.10.0.84
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pycparser==2.22
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
spandrel==0.4.0
sympy==1.13.1
tifffile==2024.9.20
tokenizers==0.20.3
tomli==2.0.2
torch==2.5.0+cu121
torchaudio==2.5.0+cu121
torchmetrics==1.5.2
torchsde==0.2.6
torchvision==0.20.0+cu121
tqdm==4.67.0
trampoline==0.1.2
transformers==4.46.2
typer==0.13.0
typing_extensions==4.12.2
urllib3==1.26.20
wrapt==1.16.0
yarl==1.17.1
